requirements.txt
ag2==0.9.3
annotated-types==0.7.0
anyio==4.9.0
asyncer==0.0.8
attrs==25.3.0
autogen==0.9.3
autogen-agentchat==0.4.9.3
autogen-core==0.4.9.3
autogen-ext==0.4.9.3
autogen-ext-mcp==0.2.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
coverage==7.9.1
diskcache==5.6.3
docker==7.1.0
fastapi==0.115.13
genson==1.3.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
json-schema-to-pydantic==0.3.0
jsonref==1.1.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
markdown-it-py==3.0.0
mcp @ git+https://github.com/modelcontextprotocol/python-sdk.git@dced22344d80d85884df9c26281fdfe0331e9c65
mdurl==0.1.2
ollama==0.5.1
opentelemetry-api==1.34.1
packaging==25.0
pillow==11.2.1
pluggy==1.6.0
protobuf==5.29.5
pydantic==2.11.7
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
pytest==8.4.1
pytest-asyncio==1.0.0
pytest-cov==6.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
RapidFuzz==3.13.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
responses==0.25.7
rich==14.0.0
rpds-py==0.25.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sse-starlette==2.3.6
starlette==0.46.2
termcolor==3.1.0
tiktoken==0.9.0
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.34.3
uvloop==0.21.0
watchdog==6.0.0
watchfiles==1.1.0
websockets==15.0.1
zipp==3.23.0
